// src/model/BrandCategoryResponseDto.tsx
export interface BrandCategoryResponseDto {
    id:             number;
    createdDate:    string;
    createdDateStr: Date;
    parentId:       number;
    name:           string;
    logoLg:         string;
    logoMd:         string;
    logoSm:         string;
    order:          null;
    featured:       null;
    level:          string;
    description:    null;
    icon:           null;
    subCategories:  null;
    brands:         null;
    fullName:       null;
}